client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "PyJWT>=2.9.0",
    "requests>=2.32.3",
]
